API Commit
This commit fixes basic authorization checks to all API endpoints

+++ api/accounts/delete/index.js
@@ -65,11 +65,31 @@ module.exports = async (request, response) => {
         return;
     }
 
+    if (session["session-status"] != "ACTIVE") {
+        response.writeHead(403, { "Content-Type": "application/json" });
+
+        response.write(JSON.stringify({ "code": 403, "heading": "Forbidden", "message": "Something went wrong!" }));
+
+        response.end();
+
+        return;
+    }
+
     account = await new Promise((resolve, reject) => {
         fs.readFile("/home/" + account + "/index.json", (error, content) => {
             return error == null ? resolve(JSON.parse(content)) : reject(new Error(error["message"]));
         });
     });
+
+    if (account["account-status"] != "ACTIVE") {
+        response.writeHead(403, { "Content-Type": "application/json" });
+
+        response.write(JSON.stringify({ "code": 403, "heading": "Forbidden", "message": "Something went wrong!" }));
+
+        response.end();
+
+        return;
+    }
 
     await new Promise((resolve, reject) => {
         fs.rm("/home/" + account["account-id"] + "/", { recursive: true }, (error) => {
